fix(task-graph): forward full run config, close iterator stream trap, wire dependency-JSON edges
- TaskGraph.run() hand-copied nine config fields into runGraph, silently
  dropping enforceEntitlements and matchAllEmptyInputs; spread the config
  so every field forwards (runPreview already did).
- A ReduceTask whose body ended in a streaming task copied the x-stream
  annotation onto its aggregate output schema, routed the run through
  IteratorTask.executeStream, and returned the input untouched without
  running an iteration. Strip x-stream from copied ports and make the
  iterator's executeStream throw a TaskConfigurationError instead of
  silently yielding its input.
- createGraphFromDependencyJSON never read item.dependencies; JsonTask
  re-wired them at the top level, but graphs recursed through subtasks lost
  their edges silently (and toDependencyJSON did not round-trip). Wire the
  dataflows inside the factory and drop JsonTask's duplicate loop.
- Document ABORTING as the real post-abort status in EXECUTION_MODEL.md
  (the ABORTED status it described does not exist).
